refactor: move to type directory
