+++ common-module/src/main/java/com/yapp/crew/domain/type/ExerciseType.java
@@ -1,4 +1,4 @@
-package com.yapp.crew.domain.status;
+package com.yapp.crew.domain.type;
 
 public enum ExerciseType {
   RUNNING("러닝"),
